feat: Add HMAC token API URL handling for Cloudflare streams and enhance download functionality
