+++ .env.default
@@ -7,6 +7,7 @@ MUSIC_PLAYLIST_API_URL=MUSIC_PLAYLIST_API_URL
 MUSIC_STREAM_API_URL=MUSIC_STREAM_API_URL
 RECENT_MUSIC_API_URL=RECENT_MUSIC_API_URL
 FAVORITE_API_URL=FAVORITE_API_URL
+HMAC_TOKEN_API_URL=HMAC_TOKEN_API_URL
 
 APP_ENV=default
 DEBUG=true
